separate out vim module
